// src/store/modules/banner/mutation-types.ts
export enum BannerMutationTypes {
  SET_INDEX_BANNER = 'SET_INDEX_BANNER',
  SET_SECKILL_BANNER = 'SET_SECKILL_BANNER',
  SET_GROUP_BANNER = 'SET_GROUP_BANNER',
  SET_FULL_MINUSS_BANNER = 'SET_FULL_MINUSS_BANNER',
  SET_NEWS_GOODS_BANNER = 'SET_NEWS_GOODS_BANNER',
  SET_SPECIAL_PRICE_BANNER = 'SET_SPECIAL_PRICE_BANNER',
  SET_RANK_BANNER = 'SET_RANK_BANNER',
  SET_INTEGRAL_BANNER = 'SET_INTEGRAL_BANNER',
  SET_IS_LOADED = 'SET_IS_LOADED',
}